poc for bootcsum based boot lookup
As it stands ostree's deployment path is unpredictable at the time the
os image is assembled as it's either dependant on future knowledge (e.g.
the final checksum of the commit) or the deployment order (e.g. boot.X
links and deployment number).

This is ok as long as that information can be passed via configuration
at deployment time (e.g. the ostree= kernel argument). However when
securing a system for verified boot one doesn't want unauthorized kernel
arguments to be added to the kernel command line. Unfortunate as a
deployed system also shouldn't have private keys to sign items, there
isn't really way in which one can have both dynamic kernel command lines
generate on the device and a way to verify them.

As a poc solution for this, give prepare-root a tad more knowledge such
that it can look up the deployment target from being given only the boot
checksum and the osname under the assumptions that:

* One only ever cares about booting the last deployment of an os image
* Each deployment has a unique boot checksum[0]
* The os is deployed under the expected osname

With this in place, a strategy can be used where is defined bootcsum is e.g. a
checksum of all the various boot artifacts (uefi stub, kernel,
initramfs, cmdline (salted with build version for uniqueness)) to be
input into a built-in cmdlines in a single image (which doesn't except
other kernel arguments).

0: Not strictly needed, but if not rollback to a previous image won't
   work

Signed-off-by: Sjoerd Simons <sjoerd@luon.net>